requirements.txt
requests
flask
faker
future
sh
docopt
pyaml
simplejson
nose
python-hostlist
prettytable
pytimeparse
cloudmesh_base
mongoengine
tabulate
cmd3
fake-factory
passlib
cloudmesh
texttable
Flask-HTTPAuth